requirements.txt
flask
gunicorn
opencv-python-headless
mediapipe
numpyProcfile
werkzeug
